util: fix event handling of certain instructions; more debug messages
For conditional branch and ret instruction, we may get an Error message
right after the entire block is retired. At this point, we have no idea
whether we should have got a DirectBranch/IndirBranch or not, as those
are optional when in BTM mode or return compression is enabled. Thus,
do not retire the block and re-sync according to the next Sync message.